pyproject.toml
[tool.poetry]
name = "pd-backend"
version = "0.1.0"
description = "REST API for interacting with the ECS PD SQL database"
authors = ["bennettscience <brian.bennett2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.25.0"
webargs = "^8.0.0"
Flask-Login = "^0.5.0"
Authlib = "^0.15.3"
pyOpenSSL = "^20.0.1"
requests = "^2.25.1"
Flask-WeasyPrint = "^0.6"
google-api-python-client = "^2.12.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
python-dotenv = "^0.17.1"
flake8 = "^3.9.2"
isort = "^5.8.0"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
